requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
distlib==0.3.9
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.26.3
wheel==0.43.0
